#1039: ARM64 versions of vamp-simple-host, dependencies, and plugins
Two plugins: QM and Segmentino.  There are now scripts that will fetch the dylibs from various places (homebrew and a custom version of vamp-simple-host), and will stick them into the right folder.  A second script makes them standalone (that is, they have no dependencies on libraries outside of the vamp-standalone directory and normal system dylibs).  This speeds up bar/beat detection and segmenting on ARM64 machines (which is all we support now).