#1383: simple indicator that Vamp is broken
At startup time, it takes only a fraction of a second to see if Vamp is there, and if so, check that it has access to beat/bar detection AND segmentation plugins.  If one or both of these are inaccessible, a ! warning triangle will appear in the status bar, just before the "Songs found" count.  Hovering over that message will indicate a vamp error code, and a request to report it.  This should only be seen if the Gatekeeper is not letting vamp run.  In my testing, it does run correctly, but that might be only on my system.  This way, failures on other systems can be reported more easily.  The real fix here is to digitally sign everything, but I've tried at least twice before, and failed both times.  I'll try again!