#1387: one release script to bind them all
A single release script calls the other ones now.  It defaults to me as release engineer, but the defaults can be overridden.  Invocation: cd to test123 in the build directory of choice, then type ./releaseSquareDesk app-specific-password .  It will make the Apple Silicon Install  version of SquareDesk, it will sign and notarize it, and then it will make the DMG file and sign that.