#1385: SD panes must have non-zero size the very first time seen
The checkers and the extra tabs views were not visible by default, upon first startup of SquareDesk on a brand new system.  Users would have a hard time knowing how to see the checkers view.  Now, both views are visible the very first time, and if the user changes them, then that size is remembered.