#1404: when playlist changes, reflect change in songTable
if songTable is showing something that would also have changed.
There's still a bug if you make a playlist change when it is NOT showing in the songTable right now, when you click on the TreeWidget to view that playlist it will show the old version until you make a change to that playlist again while it's being viewed.