#1347: fixed white lines at left end of SVGWaveformSlider
The root cause was that when tabs were switched, the view geometry needed to be rreset (don't ask me why -- I never read anything about this in the Qt documentation).  I gave this problem to Claude to solve when I couldn't find the problem after hours of debugging.  Claude did this fix in about 5 minutes total (I'm very, very impressed).