#1461: allow typing in Taminations
Not sure why this worked before at all.  But, it seems to be fixed now with this change.  If CMD-<some key> is pressed, it will NOT go to Taminations.  Note that CMD-V (paste) and CMD-Z (undo) work exactly as they do in a safari browser.  Copy and Paste buttons do not work in Taminations in SquareDesk yet.