#1462: initial scaffold for SD MCP server
This has zero SD features right now.  It's the very first step towards creating an SD MCP server in C++ to provide information to an LLM (e.g. Claude Desktop).  It exports exactly one server: "echo", that gets a text string and echoes it back to the LLM caller.  Double-click on the squaredesk-mcp.dxt to install it into Claude Desktop.