#1492: All song loads will now reset the patter timer to zero
Takes care of the case where (e.g. in a Festival or a Class) there are no singing calls, and the normal sequence is patter + patter + patter, etc.  Clicking on the darkWarningLabel still works to reset the patter timer manually, AND there's also the (new) Reset Patter Timer keyboard shortcut.